+++ Solutions/PowerShell.Commands/Commands/Branding/AddNavigationNode.cs
@@ -2,14 +2,15 @@
 using OfficeDevPnP.PowerShell.Commands.Enums;
 using System;
 using System.Management.Automation;
+using OfficeDevPnP.Core.Enums;
 
 namespace OfficeDevPnP.PowerShell.Commands
 {
     [Cmdlet(VerbsCommon.Add, "SPONavigationNode")]
     public class AddNavigationNode : SPOWebCmdlet
     {
-        [Parameter(Mandatory = true, HelpMessage="Either 'Top' or 'Quicklaunch'")]
-        public NavigationNodeType Location;
+        [Parameter(Mandatory = true)]
+        public NavigationType Location;
 
         [Parameter(Mandatory = true)]
         public string Title;
@@ -22,16 +23,16 @@ namespace OfficeDevPnP.PowerShell.Commands
 
         protected override void ExecuteCmdlet()
         {
-            if(Url == null)
+            if (Url == null)
             {
                 ClientContext.Load(SelectedWeb, w => w.Url);
                 ClientContext.ExecuteQueryRetry();
                 Url = SelectedWeb.Url;
             }
-            SelectedWeb.AddNavigationNode(Title, new Uri(Url), Header, Location == NavigationNodeType.QuickLaunch);
+            SelectedWeb.AddNavigationNode(Title, new Uri(Url), Header, Location);
         }
 
     }
 
-    
+
 }